refactor: moved CSkipSet from CreditPool.h to separate file
